created slug folder, and move relevant files to this folder
